Move leftover flags that seemed forgotten.
